% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoContradict.R
\name{AutoContradict}
\alias{AutoContradict}
\title{Function applies the Contradict function to a set of element pairs that share
  the contradictory relationships specified in contradict.link.types}
\usage{
AutoContradict(contradict.link.types, data, coh.matrix)
}
\arguments{
\item{contradict.link.types}{the subset of contradictory links in the data 
set we are seeking to model}

\item{data}{data frame with element pairs and link types between them}

\item{coh.matrix}{a relationship matrix whose weights will be modified by 
this function}
}
\value{
A relationship matrix with weights reflecting the contradictory 
  relationships specified in this function
}
\description{
Function applies the Contradict function to a set of element pairs that share
  the contradictory relationships specified in contradict.link.types
}
\examples{
AutoContradict()
}
\seealso{
\code{\link{Contradict}} which this function wraps
}

