% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoDataObs.R
\name{AutoDataObs}
\alias{AutoDataObs}
\title{Function applies the DataObs function to a subset of data as specidied by the
  data.link.types argument}
\usage{
AutoDataObs(data.link.types, coh.matrix, data, data.node.id)
}
\arguments{
\item{data.link.types}{character vector that corresponds to the subset of 
links in the dataset -links to a special observation unit- you are trying 
to model}

\item{coh.matrix}{the matrix holding the link weights}

\item{data}{the dataset specifying relationship types between elements}

\item{data.node.id}{the index in coh.matrix corresponding to the special 
observation unit}
}
\value{
A matrix with weights refecting the explanatory relationships 
  specified in the function
}
\description{
Function applies the DataObs function to a subset of data as specidied by the
  data.link.types argument
}
\examples{
AutoDataObs()
}
\seealso{
\code{\link{DataObs}} which this function wraps
}

