% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceFunctions.R
\name{AutoExplain}
\alias{AutoExplain}
\title{Function applies the explain function to a subset of data as specified by the
  explain.link.type argument}
\usage{
AutoExplain(explain.link.types, coh.matrix, data)
}
\arguments{
\item{coh.matrix}{a relationship matrix that will hold the link weights 
representing the relationships between the different elements of the 
explanatory nettwork}

\item{data}{data frame witht he arguments we are trying to model}

\item{explain.link.type}{a character vector that contains the variable names
of the links in the dataset you are making a network out of.}
}
\value{
a relationship matrix with the weights reflecting the explanatory 
  relationships specified in the function
}
\description{
Function applies the explain function to a subset of data as specified by the
  explain.link.type argument
}
\seealso{
\code{\link{Explain}} which the function wraps
}

