% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceFunctions.R
\name{CohDiff}
\alias{CohDiff}
\title{Finds the difference in activation values between nodes in a normal coherence
  network and one where the value of one of the nodes has be held at a fixed 
  activation value}
\usage{
CohDiff(coh.matrix, data.node.id, fixed.node.id, fixed.act.lvl)
}
\arguments{
\item{coh.matrix}{a amatrix holding the weights of the links in the system of 
arguments whose coherence you are trying to maximize}

\item{data.node.id}{the index of the the special observation unit}

\item{fixed.node.id}{a numeric value representing the index of the node whose
activation you are holding at a fixed value}

\item{fixed.act.val}{the value, usually -1 or 1, at which you are holding the
activation level of the node represented by fixed.node.id}
}
\value{
A vector of the difference between the activation values of the 
   propositions in coh.matrix and the activation values one gets when the 
   fix the activation value of a node in the network.
}
\description{
Finds the difference in activation values between nodes in a normal coherence
  network and one where the value of one of the nodes has be held at a fixed 
  activation value
}
\seealso{
\code{\link{Coherence}} and \code{\link{CoherenceFixed}} which this
  function wraps
}

