% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceFunction.R
\name{Coherence}
\alias{Coherence}
\title{This function updates the activation levels of nodes in neural network to 
  solve coherence problems in accordance to the ECHO algorithm presented in 
  Thagard (1989)}
\usage{
Coherence(coh.matrix, data.node.id)
}
\arguments{
\item{coh.matrix}{the relationship matrix whose coherence we are trying to 
maximize}

\item{data.node.id}{the index in coh.matrix corresponding to the special 
observation node}
}
\value{
a vector of activation values that maximizes the coherence of the 
  network
}
\description{
This function updates the activation levels of nodes in neural network to 
  solve coherence problems in accordance to the ECHO algorithm presented in 
  Thagard (1989)
}
\examples{
Coherence()
}

