% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceFunctions.R
\name{CoherenceFixed}
\alias{CoherenceFixed}
\title{A function that calculates the coherence of an explanatory system when the 
  value of a specified node is fixed}
\usage{
CoherenceFixed(coh.matrix, data.node.id, fixed.node.id, fixed.act.lvl)
}
\arguments{
\item{coh.matrix}{the relationship matrix whose coherence we are trying to 
maximize}

\item{data.node.id}{the index in coh.matrix corresponding to the special 
observation node}

\item{fixed.node.id}{the index of the element whose value you are holding at 
fixed value}

\item{fixed.act.lvl}{the value you are fixing a node at}
}
\value{
A vector of activation values corresponding to the network 
  represented by coh.matrix
}
\description{
A function that calculates the coherence of an explanatory system when the 
  value of a specified node is fixed
}
\seealso{
\code{\link{Coherence}} which this function is a variation of.
}

