% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohForceMatrix.R
\name{CohForceMatrix}
\alias{CohForceMatrix}
\title{A function that applies the CohDiff function to all nodes in a coherence 
  network}
\usage{
CohForceMatrix(coh.matrix, data.node.id, fixed.act.lvl)
}
\arguments{
\item{coh.matrix}{A matrix holding the weights of all the links in the 
coherence network}

\item{data.node.id}{A numeric index that corresponds to the special 
observation unit in the network}

\item{fixed.act.lvl}{The activation level, usually -1 or 1, at which you want
each  of the nodes to be held at one at a time}
}
\value{
A matrix where the columns are the differences between the activation
  level of each node when the element corresponding to that colum is fixed to 
  the activation value specified in fixed.act.lvl
}
\description{
A function that applies the CohDiff function to all nodes in a coherence 
  network
}
\examples{
  CohForceMatrix()
}
\seealso{
\code{\link{CohDiff}} which this function wraps
}

