% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contradict.R
\name{Contradict}
\alias{Contradict}
\title{Creates or modifies weight of link between two elements in a matrix 
  representing a Coherence network.}
\usage{
Contradict(arg1, arg2, coh.matrix = c1)
}
\arguments{
\item{arg1}{The index of the first contradictory element}

\item{arg2}{The index of the second contradictory element, must be different 
from arg1}

\item{coh.matrix}{The matrix that will contain the value of the links in the 
Network whose Coherence we are maximizing}
}
\value{
A matrix with new weight representing the contradictory relationship
  between the elements represented by arg1 and arg2.
}
\description{
Creates or modifies weight of link between two elements in a matrix 
  representing a Coherence network.
}
\examples{
Contradict()
}

