% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain.R
\name{Explain}
\alias{Explain}
\title{Creates explanatory links between elements in a matrix}
\usage{
Explain(arg1, arg2, coh.matrix = c1)
}
\arguments{
\item{arg1}{a vector of indices corresponding to propositions that co-explain
the argument represented by arg 2}

\item{arg2}{a single value corresponding to the proposition being explained}

\item{coh.matrix}{The matrix that will contain the value of the links in the 
network whose coherence we are maximizing}
}
\value{
: A matrix with new link weights representing the explanatory 
  relationships between the elements in arg1 and arg2.
}
\description{
Creates explanatory links between elements in a matrix
}
\examples{
Explain()   
}

